\docType{data}
\name{hep}
\alias{hep}
\title{Heptathlon data set}
\value{
  A data frame
}
\description{
  Results of the Olympic heptathlon competition, Seoul,
  1988. This dataset contains twenty-five observations on
  eight variables: \describe{ \item{hurdles:}{ results of
  the 100m hurdles.} \item{highjump:}{ results of the high
  jump.} \item{shot:}{results of the shot put.}
  \item{run200m:}{results of the 200m race.}
  \item{longjump:}{results of the long jump.}
  \item{javelin:}{results of the javelin.}
  \item{run800m:}{results of the 800m race.}
  \item{score:}{final score.} }
}
\examples{
data(hep)
}
\keyword{datasets}

