\docType{package}
\name{nclRcourses-package}
\alias{nclRcourses}
\alias{nclRcourses-package}
\alias{nclrcourses}
\title{R courses at Newcastle University}
\description{
  Functions and datasets used in for external R courses at
  Newcastle University
}
\author{
  \email{colin.gillespie@newcastle.ac.uk}
}
\keyword{package}

