% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nclRprogramming-package.R
\docType{package}
\name{nclRprogramming-package}
\alias{nclRprogramming}
\alias{nclRprogramming-package}
\title{R courses at Newcastle University}
\description{
Functions and datasets used in for external R courses at Newcastle University
}
\keyword{package}

