\docType{data}
\name{dummy_data}
\alias{dummy_data}
\title{Dummy data set}
\value{
  A data frame
}
\description{
  This data frame dummy_data represents an experiment,
  where we have ten treatments: A, B, ..., J and
  measurements at some time points. We want to create a
  scatter plot of measurement against time, for each
  treatment type.
}
\examples{
data(dummy_data)
}
\keyword{datasets}

